Updated main files

+++ src/main/java/MainFrame.java
@@ -2,9 +2,7 @@ import javax.swing.*;
 import javax.swing.filechooser.FileNameExtensionFilter;
 import javax.swing.filechooser.FileFilter;
 import java.awt.event.ActionEvent;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 
 public class MainFrame extends JFrame{
     private JPanel panel1;
@@ -36,24 +34,32 @@ public class MainFrame extends JFrame{
 	//  CONSTRUCTOR FOR OUR WINDOW
     public MainFrame(){
         JScrollPane scrollPane = new JScrollPane(codeEdit, JScrollPane.VERTICAL_SCROLLBAR_ALWAYS, JScrollPane.HORIZONTAL_SCROLLBAR_ALWAYS);
-
+		
+		codeEdit = new JTextArea();
 
         //  Add Filter for text files
         FileFilter txtFilter = new FileNameExtensionFilter("Plain text", "txt");
         fc.setFileFilter(txtFilter);
 
         //  MENU and MENU ITEMS
-        add(scrollPane);
-        myMenuBar = new JMenuBar();
-        setJMenuBar(myMenuBar);
-		file = new JMenu();
-        myMenuBar.add(file);
+        
+		myMenuBar = new JMenuBar();
+		
+		file = new JMenu("File");
         file.add(Open);
         file.add(Save);
-        file.add(Create);
+        file.add(New);
         file.add(Close);
-
-        setDefaultCloseOperation(EXIT_ON_CLOSE);
+		
+        myMenuBar.add(file);
+		
+		add(myMenuBar);
+        setJMenuBar(myMenuBar);
+        
+		//add(scrollPane);
+        //scrollPane.add(codeEdit);
+        add(codeEdit);
+		setDefaultCloseOperation(EXIT_ON_CLOSE);
         pack();
         setLocationRelativeTo(null);
         setVisible(true);
@@ -63,9 +69,7 @@ public class MainFrame extends JFrame{
     Action Open = new AbstractAction("Open"){
         @Override
         public void actionPerformed(ActionEvent e){
-            if(fc.showOpenDialog(null) == JFileChooser.APPROVE_OPTION){
-                openFile(fc.getSelectedFile().getAbsolutePath());
-            }
+            openFile();
         }
     };
 
@@ -76,7 +80,7 @@ public class MainFrame extends JFrame{
         }
     };
 
-    Action Create = new AbstractAction("Create") {
+    Action New = new AbstractAction("New") {
         @Override
         public void actionPerformed(ActionEvent e) {
             createFile();
@@ -92,35 +96,54 @@ public class MainFrame extends JFrame{
 
 
     //  PERFORM ACTIONS METHODS HERE:
-    public void openFile(String fileName){
-        FileReader fr = null;
-        try{
-            fr = new FileReader(fileName);
-            codeEdit.read(fr, null);
-            fr.close();
-            setTitle(fileName);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public void openFile(){
+        if(fc.showOpenDialog(null) == JFileChooser.APPROVE_OPTION){
+			File fi = new File(fc.getSelectedFile().getAbsolutePath());
+			
+			try { 
+				// String 
+				String s1 = "", sl = ""; 
+
+				// File reader 
+				FileReader fr = new FileReader(fi); 
+
+				// Buffered reader 
+				BufferedReader br = new BufferedReader(fr); 
+
+				// Initilize sl 
+				sl = br.readLine(); 
+
+				// Take the input from the file 
+				while ((s1 = br.readLine()) != null) { 
+					sl = sl + "\n" + s1; 
+				} 
+
+				// Set the text 
+				codeEdit.setText(sl); 
+			} 
+			catch (Exception evt) { 
+				JOptionPane.showMessageDialog(new JFrame(), evt.getMessage()); 
+			}
+		}
     }
 
     private void createFile() { //  COME BACK TO THIS METHOD
-        if(fc.showOpenDialog(null) == JFileChooser.APPROVE_OPTION){
-            FileWriter fw = null;
-            try{
-                fw = new FileWriter(fc.getSelectedFile().getAbsolutePath());
-                codeEdit.write(fw);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
+        codeEdit.setText("");
     }
 
     public void saveFile(){
         if(fc.showSaveDialog(null) == JFileChooser.APPROVE_OPTION){
             FileWriter fw = null;
             try{
-                fw = new FileWriter(fc.getSelectedFile().getAbsolutePath() + ".txt");
+                String fileName = fc.getSelectedFile().getAbsolutePath();
+				if(true)//TODO: Need to check for pre-existing extension in file name
+				{
+					fw = new FileWriter(fc.getSelectedFile().getAbsolutePath());
+				}
+				else
+				{
+					fw = new FileWriter(fc.getSelectedFile().getAbsolutePath() + ".txt");
+				}
                 codeEdit.write(fw);
                 fw.close();
             } catch (IOException e){
